[ALPS05114213] battery: fix kasan ke
fix kasan ke

MTK-Commit-Id: e183b68f14c3d59d4f987ce689d5368dcb80aef7

CR-Id: ALPS05114213
Feature: Fuel Gauge

+++ drivers/power/supply/mtk_gauge.h
@@ -124,7 +124,7 @@ struct gauge_hw_status {
 	/* boot status */
 	int pl_charger_status; /* for GM2.5 */
 
-	int gspare0_reg, gspare3_reg;
+	u8 gspare0_reg, gspare3_reg;
 	int rtc_invalid;
 	int is_bat_plugout;
 	int bat_plug_out_time;
